Split numina cli functionalities in modules
